% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_functions.R
\name{bubble_ci_from_tclist}
\alias{bubble_ci_from_tclist}
\title{Creating bubble CI out of TC simulated lists}
\usage{
bubble_ci_from_tclist(dflist, center_idx, alpha = 0.1, position = 1:2,
  alpha_ci = 0.05, unit_measure = "nautical mile")
}
\arguments{
\item{dflist}{List of simulated TCs}

\item{center_idx}{Index of the path that needs to be considered as a center}

\item{alpha}{Alpha level of the bubble CI}

\item{position}{Columns position of long/lat pair. Default is 1:2}

\item{alpha_ci}{Fraction for confidence interval of area estimates 
based on uniform sampling distribution.}

\item{unit_measure}{Unit of measure used for distance}
}
\value{
\item{bubble_CI_object}{A list of matrices, each with prediction and, 
if alpha was not NULL, extra column on whether the point is within 
a specific 1-alpha countour, see the \code{\link{bubbleCI}} }
\item{area}{maximum area vector for the CB in question, see 
\code{\link{error_bands_bubbleCI}} and \code{\link{max_cumulative_area}}}
\item{area_vector}{cumulative area as one steps along the CB,
\code{\link{error_bands_bubbleCI}} and \code{\link{max_cumulative_area}}}
}
\description{
This function is a wrapper to determine CI bubble for list of simulated TCs

\strong{Warning: this function does lat and lon in reverse. This 
goes further back, and goes forward to effection the graphing function
\code{\link{ggvis_bubble_data}}.}
}
