% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_functions.R
\name{bubbleCI}
\alias{bubbleCI}
\title{Estimating the CI around a given path and a given level with symmetrical 
radius}
\usage{
bubbleCI(dflist, center_idx, alpha = 0.1, position = 1:2,
  output_length = "nautical mile")
}
\arguments{
\item{dflist}{List of simulated TCs}

\item{center_idx}{Index of the path that needs to be considered as a center}

\item{alpha}{Alpha level of the confidence interval}

\item{position}{Columns position of long/lat pair. Default is 1:2}

\item{output_length}{Unit of measure of the output}
}
\value{
A list of matrices, each with prediction and, if alpha was not 
NULL, extra column on whether the point is within a specific 1-alpha 
countour.
}
\description{
This function is a wrapper around rearrange_dflist_bubble and 
create_CI_bubble_step_track. First rearranges the dflist of simulated TC and 
then calculates the 1-alpha level bubble for each of the track points of the 
"central" TC.
}
