% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_functions.R
\name{check_points_in_bubbleCI}
\alias{check_points_in_bubbleCI}
\title{Check if points are inside bubble CI}
\usage{
check_points_in_bubbleCI(df_points, center_df, radius_df, position = 1:2)
}
\arguments{
\item{df_points}{Dataframe of points which inclusion needs to be calculated}

\item{center_df}{Dataframe of central points of the CI bubble}

\item{radius_df}{Vector of radius of the CI bubble - cast as dataframe}

\item{position}{Columns position of long/lat pair. Default is 1:2}
}
\value{
Boolean vectors on whether the points are in any of the bubble CIs 
or not
}
\description{
Given a dataframe of points this function returns a vector of the same
length with 1 or 0 whether the respective points is included in any of the CI
bubble.

NOTE: this approach has been superseded by 
\code{\link{check_points_within_diff_radius}}
}
