% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_visual_functions.R
\name{contour_list_to_df}
\alias{contour_list_to_df}
\title{Title}
\usage{
contour_list_to_df(contour_list)
}
\arguments{
\item{contour_list}{unnamed list of contour elements, of 
\itemize{
\item level The contour level.
\item x The x-coordinates of the contour.
\item y The y-coordinates of the contour.
}

This list has the same format as the \code{contour} element generated by
\code{\link{kde_from_tclist}}}
}
\value{
a list of data frames (\code{x}, \code{y}, and \code{level} columns)
}
\description{
Title
}
