% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_source.R
\name{convert_lat}
\alias{convert_lat}
\title{Convert latitude hemispheres to pos/neg}
\usage{
convert_lat(df)
}
\arguments{
\item{df}{TC data frame, where lat column has N/S at end}
}
\value{
Latitude column, converted to numeric
}
\description{
If latitude ends in "N", drop "N" and convert latitude to 
numeric. If latitude ends in "S", drop "S" and convert latitude to negative 
numeric.
}
