% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{data_sanitize}
\alias{data_sanitize}
\title{Sanitize TC data for analysis}
\usage{
data_sanitize(df)
}
\arguments{
\item{df}{Original TC data frame}
}
\value{
Sanitized TC data frame, w/ six-hour latitude/longitude observations
}
\description{
This function takes in an original TC data frame. It saves
only the rows on the six-hour marks and checks that all sequential 
observations are six hours apart. Then it returns a new data frame containing
only the latitude and longitude columns.
}
