% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_ball.R
\name{delta_structure}
\alias{delta_structure}
\title{Performs delta ball approach}
\usage{
delta_structure(data_list, alpha, dist_mat = NULL, data_deep_points = NULL,
  position = 1:2, depth_vector = NULL, area_ci_n = 2000,
  area_ci_alpha = 0.05, verbose = FALSE, ...)
}
\arguments{
\item{data_list}{list of hurricanes}

\item{alpha}{for prediction band (related to depth). Takes value in (0, 1.0), 
for a 95\% PB, set alpha to .05.}

\item{dist_mat}{distance matrix (otherwise is calculated)}

\item{data_deep_points}{data deep points from depth function 
(otherwise calculated)}

\item{position}{Columns position of long/lat pair}

\item{depth_vector}{Vector with depth values}

\item{area_ci_n}{number of observations to estimate area of covering}

\item{area_ci_alpha}{alpha level for confidence interval of area of covering}

\item{verbose}{if the distance matrix is verbose}

\item{...}{other parameters in distance calculation through 
\code{\link{distMatrixPath_innersq}}}
}
\value{
\describe{
\item{structure}{data frame of non-ordered lines of contour}
\item{area}{estimated error of contour area (float)}
\item{area_ci}{area confidence interval (vector)}
\item{delta}{optimal delta for covering}
}
}
\description{
Performs delta ball approach
}
