% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth_function.R
\name{depth_curves_to_points}
\alias{depth_curves_to_points}
\title{Get deepest curves' points in a data frame}
\usage{
depth_curves_to_points(data_list, alpha, dist_mat = NULL, position = 1:2,
  depth_vector = NULL, verbose = FALSE, ...)
}
\arguments{
\item{data_list}{list of hurricanes}

\item{alpha}{for prediction band (related to depth). Takes value in (0, 1.0), 
for a 95\% PB, set alpha to .05.}

\item{dist_mat}{distance matrix (otherwise is calculated)}

\item{position}{only needed if created 13 point reduction}

\item{depth_vector}{vector of depth values (otherwise calculated)}

\item{verbose}{if the distance matrix is verbose}

\item{...}{other parameters in distance calculation through 
\code{\link{distMatrixPath_innersq}}}
}
\value{
data frame with points in desired curves
}
\description{
Get deepest curves' points in a data frame
}
