% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_functions.R
\name{distance_mat_eulid}
\alias{distance_mat_eulid}
\title{Euclidean distance wrapper}
\usage{
distance_mat_eulid(projection_locs)
}
\arguments{
\item{projection_locs}{locations of TC in projected space (n x d)}
}
\value{
euclidean distance matrix comparing between points
}
\description{
Wrapper for calculating euclidean distance, for a single TC
}
