% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_functions.R
\name{error_bands_bubbleCI}
\alias{error_bands_bubbleCI}
\title{Estimating uncertainty of bubble CI}
\usage{
error_bands_bubbleCI(bubble_steps_CI, long_col = 1, lat_col = 2,
  unit_measure = "nautical mile")
}
\arguments{
\item{bubble_steps_CI}{list of dataframes, with each dataframe representing 
one step in the track of the most central TC}

\item{long_col}{Column index of the longitude}

\item{lat_col}{Column index of the latitude}

\item{unit_measure}{Unit measure of the distances}
}
\value{
\item{centers}{2 x n data frame of the center path used for the CB} 
\item{radius}{n length vector of radius around center path for each step}
\item{positive, negative}{2 x n data frames holding points \eqn{r} distance
away for the center point tangent to the path of the center path}
}
\description{
This function takes in a list of dataframes, representing the different CI 
bubbles built on the track of the most "central" points, and extract center 
points, radiuses and extreme of the CI bubbles per each step
}
