% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sca_estimation_and_visualization.R
\name{estimate_p_wrapper}
\alias{estimate_p_wrapper}
\title{Wrapper for getting structure of SCA lower dimensional projection}
\usage{
estimate_p_wrapper(training_structure_estimating_p, D_test, kdensity = 10)
}
\arguments{
\item{training_structure_estimating_p}{list from 
\code{\link{training_structure_estimating_p}}}

\item{D_test}{Distance matrix between test points m and training points n
(a m x n matrix)}

\item{kdensity}{k for knn density caculation in projection space}
}
\value{
\item{new_sigma_k}{distance to Kth neighbor for test set  (of points in 
training set), note this K is defined within 
\code{\link{training_structure_estimating_p}} parameter}
\item{P_test}{transition matrix after local scaling}
\item{test_projected}{data frame of projected test points}
\item{p_estimate_test}{vector of probability associated with each test point
(using kdensity parameter)}
}
\description{
Wrapper for getting structure of SCA lower dimensional projection
}
