% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde_functions.R
\name{extract_countour}
\alias{extract_countour}
\title{Selection of specific contour level from KDE object}
\usage{
extract_countour(kde_obj, alpha)
}
\arguments{
\item{kde_obj}{kde object}

\item{alpha}{contour level which needs to be extracted. Scalar in the sequence
from .01 to .99 by .01. A value of .05 gives a contour level 
  associated with the cumulative prob of 95% (and 95% confidence interval) -
  i.e. a 1 - alpha confidence contour.}
}
\value{
List of contours at (100-level) for the kde object - there is a
contour for each disconnected contour that forms the (100-level) contour
level.
}
\description{
This function extracts a specific level of contour from a kde object.

This function only works for some levels that the kde object calculated 
originally (associated with the constraints in the \code{alpha} parameter).
}
\details{
This function (renamed as \code{extract_contour}) is shared with
\pkg{timeternR} on github:
\href{https://github.com/skgallagher/timeternR}{timeternR}.
}
\examples{
\dontrun{
set.seed(8192)

x1 <- 2^rnorm(100)
y1 <- rnorm(100)
dfmat <- cbind(x1,y1)
kde_object <- ks::kde(dfmat)

cont <- extract_countour(kde_object, .05)
}
}
