% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde_functions.R
\name{fit_kde_object}
\alias{fit_kde_object}
\title{Fit kernel density estimator object}
\usage{
fit_kde_object(dfmat, h_band = NULL, position = 1:2, grid_size = rep(1000,
  2))
}
\arguments{
\item{dfmat}{TC dataframe}

\item{h_band}{optional argument for the bandwidth of the kde object. If NULL,
the optimal band would be selected through the \code{\link[ks]{kde}} 
function. Default is NULL.}

\item{position}{Columns position of long/lat pair. Default is 1:2}

\item{grid_size}{size of the grid which is going to be used for the 
evaluation of kde object. Can be reduced to speed-up computation.}
}
\value{
KDE object fitted to the dfmat dataframe.
}
\description{
This function fits a kernel density estimator to a TC dataframe, in which 
each point of the dataframe is considered to be a TC observation.
}
\examples{
\dontrun{
set.seed(8192)
x <- 2^rnorm(100)
y <- rnorm(100)
dfmat <- cbind(x,y)

kde_object <- fit_kde_object(dfmat)
}
}
