% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_functions.R
\name{get_area_diff_radius}
\alias{get_area_diff_radius}
\title{Estimate area of union of balls with different radius(wrapper)}
\usage{
get_area_diff_radius(tc_bubble_structure, n = 10000, alpha = 0.05,
  unit_measure = "nautical mile", verbose = TRUE)
}
\arguments{
\item{tc_bubble_structure}{Bubble CI list with centers, radius, positive 
and negative parts (at least a list of centers data frame and radius vector)}

\item{n}{number of points drawn uniformly within a box around the true data}

\item{alpha}{alpha level for 2 sided confidence interval estimate}

\item{unit_measure}{string of measurement unit of points, see options in
datamart::uconv.}

\item{verbose}{boolean logic if should have print outs (specifically that we
switch the columns due to longlat boolean)}
}
\value{
\item{area}{estimated area of union of balls}
\item{area_ci}{vector with lower and upper confidence interval estimate}
}
\description{
Wrapper function to estimate the area of union of balls with 
different radius (for pointwise CI), using uniform draws
}
