% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_functions.R
\name{get_area_diff_radius_inner}
\alias{get_area_diff_radius_inner}
\title{Inner function to get approximate area/ proportion of observations in for 
pointwise bubble CB}
\usage{
get_area_diff_radius_inner(tc_bubble_structure, query, size = 1,
  position = 1:2, alpha = 0.05, unit_measure = "nautical mile",
  verbose = TRUE)
}
\arguments{
\item{tc_bubble_structure}{Bubble CI list with centers, radius, positive 
and negative parts (at least a list of centers data frame and radius vector)}

\item{query}{points to check relative to the center points}

\item{size}{size of space where query points were drawn (note if size = 1, 
then you are getting the proportion of points with the balls)}

\item{position}{Columns position of long/lat pair. Default is 1:2}

\item{alpha}{alpha level for 2 sided confidence interval estimate}

\item{unit_measure}{string of measurement unit of points, see options in
datamart::uconv.}

\item{verbose}{boolean logic if should have print outs (specifically that we
switch the columns due to longlat boolean)}
}
\value{
\item{area}{estimated area of union of balls}
\item{area_ci}{vector with lower and upper confidence interval estimate}
\item{in_vec}{vector of 0/1 to indicate if the query point was inside at 
least one ball}
}
\description{
Inner calculation of proportion of points close to the center of
the pointwise bubble based CB relative to a changing radius. This function 
works in a similar way as \code{\link{get_area_inner}} created for the 
\eqn{\delta}-ball based approach.
}
\details{
This functions uses as inputs the output of 
\code{\link{error_bands_bubbleCI}}.
}
