% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{get_bearing}
\alias{get_bearing}
\title{Compute bearing and change in bearing}
\usage{
get_bearing(lat, long)
}
\arguments{
\item{lat}{Latitude vector}

\item{long}{Longitude vector}
}
\value{
List of bearing, bearing_change, and bearing_prev vectors.
\item{bearing}{where \code{bearing[i]} = Bearing between 
(\code{lat[i], long[i]}) and (\code{lat[i+1], long[i+1]})}
\item{bearing_change}{where 
\code{bearing_change[i] = bearing[i] - bearing[i-1]},
with corrections to ensure -180 < bearing_change <= 180}
\item{bearing_prev}{where \code{bearing_prev[i] = bearing[i-1]}}
}
\description{
Compute bearing and change in bearing between all consecutive 
(latitude, longitude) pairs. Also determine previous bearing at each
observation.
}
\details{
At position i, the bearing (bear) is the bearing 
between (lat[i], long[i]) and (lat[i+1], long[i+1]). Bearing is calculated 
using geodesic path (shortest path on ellipsoid).
Formula for bearing on 
\url{http://www.movable-type.co.uk/scripts/latlong.html}
}
