% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{get_block}
\alias{get_block}
\title{Determine block of TC observation}
\usage{
get_block(long, lat, east_west_prev)
}
\arguments{
\item{long}{Longitude vector}

\item{lat}{Latitude vector}

\item{east_west_prev}{"E"/"W" vector, indicating whether TC traveled 
east or west from previous observation}
}
\value{
Vector of blocks of TC observations
}
\description{
Determine block of TC observation
}
\details{
For an individual observation of a TC, block is determined by
longitude, latitude, and whether TC traveled east or west from previous
observation. Blocks were determined by creating 10 degree latitude by
10 degree longitude regions and combining regions with few observations.
}
