% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{get_east_west}
\alias{get_east_west}
\title{Determine whether bearing indicates east or west}
\usage{
get_east_west(bearing)
}
\arguments{
\item{bearing}{Vector of bearings, where \code{bearing_prev[i]} is the 
bearing between (\code{lat[i], long[i]}) and (\code{lat[i+1], long[i+1]}).
It should be included between 0 and 360 degrees.}
}
\value{
\code{eastwest[i]} stores "E" if \code{bearing[i]} indicates eastward
travel and "W" if \code{bearing[i]} indicates westward travel
}
\description{
If 0 <= bearing < 180, then bearing indicates eastward travel.
If 180 <= bearing < 360, then bearing indicates westward travel.
}
