% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{get_reg_df}
\alias{get_reg_df}
\title{Appends regression variables}
\usage{
get_reg_df(df, auto = T)
}
\arguments{
\item{df}{TC data frame}

\item{auto}{Boolean for whether to append lag change in bearing/speed
columns (for autoregressive models)}
}
\value{
TC data frame, including columns for performing bearing, speed,
and death regressions
}
\description{
Appends columns used in bearing, speed, and death regressions. 
These columns are \code{lat}, \code{long}, \code{bearing}, 
\code{bearing_change}, \code{bearing_prev}, \code{east_west_prev}, 
\code{speed}, \code{speed_change}, \code{speed_prev},
\code{timestep}, \code{death}, and \code{block}. Also includes 
\code{bearing_change_lag1} and \code{speed_change_lag1} columns for 
autoregressive models (\code{auto = T}).
}
