% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_visual_functions.R
\name{ggvis_bubble_data_centers_inner}
\alias{ggvis_bubble_data_centers_inner}
\title{Inner Plot centers for bubble data}
\usage{
ggvis_bubble_data_centers_inner(bubble_plot_data, base_graph = NULL,
  color = "pink", zoom = 4, ...)
}
\arguments{
\item{bubble_plot_data}{list of center points, lower and upper tangential 
points (each a n x 2 data frame)}

\item{base_graph}{plot to add points to}

\item{color}{color of points}

\item{zoom}{map zoom for ggmap}

\item{...}{Interior ggpath parameters}
}
\value{
ggplot object with points on it
}
\description{
Inner Plot centers for bubble data
}
