% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_visual_functions.R
\name{ggvis_paths}
\alias{ggvis_paths}
\title{Visualize TC Paths (with \code{ggplot})}
\usage{
ggvis_paths(data_out, zoom = 4, base_graph = NULL, alpha = 0.01, ...)
}
\arguments{
\item{data_out}{data frame with correct path information same as outputted 
from \code{\link{data_plot_paths_basic}} functions. If supplied with a 
list of data frames of paths of TCs (like inputted into 
\code{\link{data_plot_paths_basic}}), then the function will apply
\code{\link{data_plot_paths_basic}} first. \emph{Note: will assume
"\code{position = 1:2}" unless altered in \code{...}, see 
documentation for \code{\link{data_plot_paths_basic}} for details.}}

\item{zoom}{map zoom for ggmap}

\item{base_graph}{ggplot object for base graph 
(created from data_out otherwise)}

\item{alpha}{opacity of curves}

\item{...}{can pass \code{position} values if desirable}
}
\value{
ggplot visualisation of the curve
}
\description{
Visualize TC Paths (with \code{ggplot})
}
