% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sca_estimation_and_visualization.R
\name{inner_sca_projection}
\alias{inner_sca_projection}
\title{Inner Function to apply SCA projection of a set of simulated curves and 
captures the associated probability values in the lower dimensional space}
\usage{
inner_sca_projection(test_data, train_list_13_point, train_info,
  position = 1:2, verbose = TRUE)
}
\arguments{
\item{test_data}{list of paths of test curves}

\item{train_list_13_point}{list of 13 point compression of the paths of 
training curves, each in longlat format for the first 2 columns}

\item{train_info}{structure for SCA algorithm from training data outputted
from \code{\link{training_structure_estimating_p}}.}

\item{position}{position of test_data's longitude and latitude columns 
respectively}

\item{verbose}{if TRUE each step of the process is expressed (1) 13 point 
compression, (2) creation of distance from test to training matrix, and 
(3) estimation of lower dimension projection and associated probabilites.}
}
\value{
\item{projection}{data frame of associated projected points of test / 
simulated curves}
\item{p_estimate}{probability estimates in the lower dimensional space of the
test / simulated curves}
}
\description{
Inner Function to apply SCA projection of a set of simulated curves and 
captures the associated probability values in the lower dimensional space
}
