% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creating_curves.R
\name{is_dead_inner}
\alias{is_dead_inner}
\title{Simulate whether to stop growing TC}
\usage{
is_dead_inner(path, death_regs, bad_locations, death_rate)
}
\arguments{
\item{path}{TC data frame, containing obs thus far in simulation}

\item{death_regs}{List of block-specific death regressions}

\item{bad_locations}{Names of blocks containing <= 1 TC death}

\item{death_rate}{1 / mean length of TCs in train data}
}
\value{
Boolean for whether to stop growing TC
}
\description{
For regression approach to TC deaths, this function simulates
whether to stop growing TC.
}
\details{
If using block-specific death regressions to determine TC death, 
this function determines whether to stop growing TC. If block contains >= 2 
TC deaths, most recent observation characteristics are plugged into
block-specific regression to predict death probability. If block contains <= 1 
TC death, predicted death probability is 1 / mean length of training TCs.
(This is the MLE under an exponential distributional assumption on TC length.)
Decision of whether to stop TC depends on a draw from a Bernoulli distribution
with probability equal to the predicted death probability.
}
