% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde_functions.R
\name{kde_contour_area}
\alias{kde_contour_area}
\title{Area calculation of a kde contour object}
\usage{
kde_contour_area(cont)
}
\arguments{
\item{cont}{kde contour object.}
}
\value{
Area of the kde contour object.
}
\description{
This function calculates the area of a contour object by creating a closed
polygon and calculating the area of it.
}
\examples{
\dontrun{
set.seed(8192)

x1 <- 2^rnorm(100)
y1 <- rnorm(100)
dfmat <- cbind(x1,y1)
kde_object <- ks::kde(dfmat)
cont <- with(kde_object, contourLines(x = eval.points[[1]],
                                   y = eval.points[[2]],
                                   z = estimate,levels = cont["5\%"])[[1]])

cont_area <- kde_contour_area(cont)
}
}
