% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loocv_functions.R
\name{loocv_wrapper}
\alias{loocv_wrapper}
\title{LOOCV error rate}
\usage{
loocv_wrapper(path_mat_list, Dmat, K = 7, t = 1,
  output_length = "nautical mile", longlat = TRUE, project_size = 4,
  verbose = TRUE)
}
\arguments{
\item{path_mat_list}{list of paths to train on}

\item{Dmat}{distance matrix of the paths to train on}

\item{K}{K nearest neighbors value (for localization of distance matrix)}

\item{t}{power to raise the transition matrix P}

\item{output_length}{string of distance between points 
(set at "nautical mile" )}

\item{longlat}{boolean if order of columns is longlat}

\item{project_size}{dimension of projected space}

\item{verbose}{it progress bar is provided}
}
\value{
\item{Dmat}{distance matrix of the paths to train on that was inputted}
\item{predicted_curves}{list of predicted curves}
\item{diff_curves}{distance between true curves and predicted curves}
\item{K}{K nearest neighbors value (for localization of distance matrix)}
\item{t}{power to raise the transition matrix P (same as inputted)}
}
\description{
Calculates the LOOCV error rate for a specific local smoothing 
of distance \eqn{K} and the power we raise the transition matrix to \eqn{t}
}
