% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_functions.R
\name{max_cumulative_area}
\alias{max_cumulative_area}
\title{Estimating Area of the Bubble CI}
\usage{
max_cumulative_area(tc_bubble_structure)
}
\arguments{
\item{tc_bubble_structure}{Bubble CI list with centers, radius, positive 
and negative parts}
}
\value{
\item{area_vec}{area of the Bubble CI for each step}
\item{max}{maximum of the area_vec values}
}
\description{
Finds area of bubble CI, for each step, then takes the maximum
}
\details{
This functions uses as inputs the output of 
\code{\link{error_bands_bubbleCI}}.

NOTE: this approach has been superseded by \code{\link{get_area_diff_radius}}
}
