% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde_functions.R
\name{points_in_contour}
\alias{points_in_contour}
\title{Identification of points inside a contour}
\usage{
points_in_contour(cont, predict_mat, long = 1, lat = 2)
}
\arguments{
\item{cont}{kde contour object.}

\item{predict_mat}{matrix with the points to be determined in terms of
position with respect to the contour.}

\item{long}{index of the column referring to "longitude". Default is 1.}

\item{lat}{index of the column referring to "latitude". Default is 2.}
}
\value{
Vector of binary values, 1 if the point is inside the contour, 
0 is not. Dimensionality is the same as the number of rows in predict_mat
}
\description{
This function calculates whether a series of points are inside a given
contour or not.
}
\examples{
\dontrun{
set.seed(8192)

x1 <- 2^rnorm(100)
y1 <- rnorm(100)
dfmat <- cbind(x1,y1)
kde_object <- ks::kde(dfmat)
cont <- with(kde_object, contourLines(x = eval.points[[1]],
                                   y = eval.points[[2]],
                                   z = estimate,levels = cont["5\%"])[[1]])
x1 <- 2^rnorm(100)
y1 <- rnorm(100)
predict_mat <- cbind(x1,y1)

position_wrt_contour <- points_in_contour(cont, predict_mat)
}
}
