% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde_functions.R
\name{predict_kde_object}
\alias{predict_kde_object}
\title{Evaluation of points with respect to KDE object}
\usage{
predict_kde_object(kde_obj, predict_mat, alpha = NULL, position = 1:2)
}
\arguments{
\item{kde_obj}{kde object - in our case based on TC points 
(from \code{\link[ks]{kde}} )}

\item{predict_mat}{matrix with points to be predicted through KDE}

\item{alpha}{alpha level of the contour plot. Default is NULL. If not 
NULL, then an extra column will be added, in which 1 means that the value is 
above the alpha contour - i.e. within that probability contour - 
else 0 is returned.}

\item{position}{Columns position of long/lat pair. Default is 1:2}
}
\value{
Matrix with prediction and, if alpha was not NULL, extra column
on whether the point is within a specific 1-alpha countour.
}
\description{
This function evaluates a matrix of geographical points with respect to a KDE
object.
}
\examples{
\dontrun{
set.seed(8192)

x1 <- 2^rnorm(100)
y1 <- rnorm(100)
dfmat <- cbind(x1,y1)
kde_object <- ks::kde(dfmat)

x1 <- 2^rnorm(100)
y1 <- rnorm(100)
predict_mat <- cbind(x1,y1)

out_mat <- predict_kde_object(kde_object, predict_mat)
}
}
