% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_interval_pipeline_functions.R
\name{prediction_interval_pipeline}
\alias{prediction_interval_pipeline}
\title{Extract CI from the full list of simulated and true TCs}
\usage{
prediction_interval_pipeline(tc_full_sim_list, tc_true_path_list, alpha = 0.1,
  start_idx = NULL, end_idx = NULL, position = 1:2,
  unit_measure = "nautical mile", verbose = TRUE,
  kde_grid_size = rep(1000, 2), alpha_ci = 0.05,
  curve_type_vec = c("Auto_DeathRegs", "Auto_NoDeathRegs", "NoAuto_DeathRegs",
  "NoAuto_NoDeathRegs"))
}
\arguments{
\item{tc_full_sim_list}{List of simulated TCs}

\item{tc_true_path_list}{Dataframe with the true TC path to be tested}

\item{alpha}{Alpha level of all the CI}

\item{start_idx}{Starting index when selecting which TC to be simulated - if NULL, defaulted
to 1}

\item{end_idx}{Ending index when selecting which TC to be simulated - if NULL, defaulted to
the number of TCs available}

\item{position}{Columns position of long/lat pair. Default is 1:2}

\item{unit_measure}{Unit of measure used for distance}

\item{verbose}{if TRUE, an update on the process will be printed}

\item{kde_grid_size}{Two dimensional vector for the KDE grid size}

\item{alpha_ci}{Fraction for confidence interval of area estimates based on uniform
sampling distribution.}

\item{curve_type_vec}{vector with the strings referring to the different types of curves}
}
\value{
A list in which each sublist is a curve type and each sublist has for each element
a TC with the 4 different calculated prediction intervals.
}
\description{
This function iterates through different type of curves and all the simulated curves
in order to obtain the prediction intervals from the 4 different methods
}
\details{
This function is just an iterative wrapper around the function prediction_interval_single_tc
}
