% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_functions.R
\name{rearrange_dflist_bubble}
\alias{rearrange_dflist_bubble}
\title{Rearranging list of simulated TCs}
\usage{
rearrange_dflist_bubble(dflist, center_idx, long = 1, lat = 2)
}
\arguments{
\item{dflist}{List of simulated TCs}

\item{center_idx}{Index of the path that needs to be considered as a center}

\item{long}{Column index of the longitude}

\item{lat}{Column index of the latitude}
}
\value{
List with same number of elements as the points in the track of the 
"central" TC. Each element j in the list is a dataframe with the j-th points 
of the simulated TCs.
}
\description{
This function takes as an input the list of the simulated TCs, with a path 
being a dataframe of (k x 2), where k is the number of points of the path. 
The function selects the "center" TC in order to get the length of this path 
k_0. This functions creates a list of k_0 elements, in which longitude and 
latitudes are saved in first and second column respectively.
}
