% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_output_pipeline}
\alias{sample_output_pipeline}
\title{Output of PB pipeline for sample TC}
\format{list with 1 element, which name is "Auto_NoDeathRegs", which in turns
has 5 elements, "kde", "bubble_ci", "delta_ball", "convex_hull",
"depth_vector" and "time", which are the result of the PB pipeline run with
\code{sample_tc} as true TC and its simulations \code{sample_sim}.}
\usage{
sample_output_pipeline
}
\description{
This object is the result of running the PB papeline on the sample TC 
"AL032009", using the 350 simulations, which generated 350 Autoregressive
simulations with Kernel Death modelling for TC lysis.
}
\keyword{datasets}
