% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_sim}
\alias{sample_sim}
\title{Example List of 350 Simulated Curves, AL032009}
\format{A list of simulated curves, each element is a matrix with format:
\describe{
  \item{long}{longitude of the eye of the simulated curve}
  \item{lat}{latitude of the eye of the simulated curve}
}}
\usage{
sample_sim
}
\description{
A list of simulated curvesx, each element in the list is matrix
(with different length), given the format below.
 
Created by using the first 3 points of \code{sample_tc}, with
name AL032009.
}
\details{
This simulation was created using the Autoregressive change in bearing and
speed models with Kernel-based Lysis (death) models.
}
\keyword{datasets}
