% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_tc}
\alias{sample_tc}
\title{Example TC, AL032009}
\format{A data frame with 69 rows and  variables:
\describe{
  \item{date}{Date time step recorded (format "YYYYmmdd")}
  \item{time}{Time of day time step was recorded, 
  (military time, e.g. 600, 1200, 1600, 1800, 0) }
  ...
  \item{lat}{latitude of the eye of the storm}
  \item{long}{longitude of the eye of the storm}
  ...
}}
\source{
\url{http://www.aoml.noaa.gov/hrd/hurdat/hurdat2-1851-2015-070616.txt}
\url{http://www.aoml.noaa.gov/hrd/hurdat/Data_Storm.html}
}
\usage{
sample_tc
}
\description{
A data frame of all recorded time steps of a single TC, a slightly cleaned 
version than that provided by the National Oceanic and Atmospheric 
Administration's data set.
}
\keyword{datasets}
