% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_tc_name}
\alias{sample_tc_name}
\title{Name of sample TC, AL032009}
\format{string with TC name associated with \code{sample_tc} and the 
true TC who's first 3 points were used to create the simulated curves 
in \code{sample_sim}.}
\usage{
sample_tc_name
}
\description{
"AL032009" means that the TC was in the Atlantic Basin (AL), the 3rd TC that
 year (03), in year 2009 (2009).
}
\keyword{datasets}
