% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth_function.R
\name{selected_paths_to_df}
\alias{selected_paths_to_df}
\title{Create a data frame of points in selected curves}
\usage{
selected_paths_to_df(data_list, desired_index = NULL, verbose = TRUE)
}
\arguments{
\item{data_list}{list of hurricanes}

\item{desired_index}{which hurricanes to be included}

\item{verbose}{if progress is to be reported in creation of data frame}
}
\value{
data frame with points in desired curves
}
\description{
Create a data frame of points in selected curves
}
