% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_ball.R
\name{steps_along_2d_line}
\alias{steps_along_2d_line}
\title{Create n equidistance points}
\usage{
steps_along_2d_line(line, n_steps = 1000)
}
\arguments{
\item{line}{2 x 2 matrix of edge points of line}

\item{n_steps}{integer number of steps (n)}
}
\value{
n x 2 matrix with points on path
}
\description{
Inner Function to create n equidistant points along a line
}
