% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_functions.R
\name{swap2DfCols}
\alias{swap2DfCols}
\title{Swap 2 columns}
\usage{
swap2DfCols(data_df)
}
\arguments{
\item{data_df}{data frame (assumed to have 2 columns)}
}
\value{
First 2 columns of the data-frame, with their position reversed.
}
\description{
Swap the 2 columns in a Data Frame

NOTE: not sure this should be exported
}
