% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thirteen_point_compression.R
\name{thirteen_points}
\alias{thirteen_points}
\title{Thirteen point compression}
\usage{
thirteen_points(df2, longlat = TRUE, output_length = "nautical mile")
}
\arguments{
\item{df2}{2 column data frame with all points lat/long}

\item{longlat}{boolean if the order of the columns is lat/long 
(will be reverse them if FALSE)}

\item{output_length}{string of distance between points 
(set at "nautical mile" )}
}
\value{
new_13compression data.frame (13 x 2) of points along the path 
equally spaced
}
\description{
Coverts list of point locations to 13 equally spaced points
}
\examples{
\dontrun{
df2 <-data.frame(matrix(
                 c(27.5, -87.5,
                 27.6, -87.9,
                 27.8, -88.2,
                 28.2, -88.4,
                 28.6, -88.6,
                 29.3, -89.0,
                 29.7, -89.4,
                 29.9, -89.5,
                 30.6, -90.4,
                 31.1, -91.6,
                 31.5, -92.8,
                 31.8, -93.8,
                 32.2, -94.7,
                 32.7, -95.0),ncol=2,byrow=T))
library(rworldmap)
newmap <- getMap(resolution = "low")
plot(newmap, ylim = c(10, 47), xlim = c(-90, -10), asp = 1)

lines(df2[,2], df2[,1], col = "red")
news_df2 <- thirteen_points(df2, longlat = F)
points(news_df2[,1], news_df2[,2], col = "red")
}
}
