% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sca_estimation_and_visualization.R
\name{training_structure_estimating_p}
\alias{training_structure_estimating_p}
\title{Creates structure for SCA algorithm from training data}
\usage{
training_structure_estimating_p(D_train, K = 4, t = 1, dim = 5,
  plot_n = 0)
}
\arguments{
\item{D_train}{Distance Matrix for training points}

\item{K}{K nearest neighbors value (for localization of distance matrix)}

\item{t}{power to raise the transition matrix P}

\item{dim}{dimension of projected space}

\item{plot_n}{number of original eigenvectors to plot (if any)}
}
\value{
\item{old_sigma_k}{distance to Kth nearest neighbor for each training point}
\item{P_train}{transition matrix of training data after local scaling}
\item{phi_map_out}{right eigenvector decomposition taking into account K, t}
\item{train_projected}{data frame of training points in projected space}
\item{structure}{parameters list (k, t, dim from inputted parameters)}
}
\description{
Creates structure for SCA algorithm from training data
}
