% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_ball.R
\name{delta_ball_wrapper}
\alias{delta_ball_wrapper}
\title{Run delta ball analysis}
\usage{
delta_ball_wrapper(data_raw, n_steps = 1000, remove_duplicates = F)
}
\arguments{
\item{data_raw}{data frame with center points of balls}

\item{n_steps}{number of equidistance points along the line, past delta 
on both sides, that will be checked to approximate all points along the line}

\item{remove_duplicates}{boolean if need to remove duplicates in data_raw}
}
\value{
data frame of exterior lines (not ordered)
}
\description{
Run delta ball analysis and get outline of points. Runs all 
analyses to get dataframe with regular lines. Part of the approach in 
"Computing Polygonal Surfaces from Unions of Balls" by Tam and Heidrich.
}
\details{
\strong{See comment at bottom of function: this function I think does the 
incorrect alignment - effects all the way to \code{ggvis_delta_ball_contour}}
}
