% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth_function.R
\name{depth_function}
\alias{depth_function}
\title{Depth calculation}
\usage{
depth_function(dist_matrix)
}
\arguments{
\item{dist_matrix}{a n x n square positive symmetric matrix}
}
\value{
depth vector length n with depth values associated with indices in 
dist_matrix
}
\description{
calculates a depth vector using a distance matrix
}
