% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_visual_functions.R
\name{ggvis_paths}
\alias{ggvis_paths}
\title{Visualize TC Paths (with ggplot)}
\usage{
ggvis_paths(data_out, zoom = 4, base_graph = NULL, alpha = 0.01)
}
\arguments{
\item{data_out}{data frame with correct path information same as outputed 
from data_plot_sc_paths functions}

\item{zoom}{map zoom for ggmap}

\item{base_graph}{ggplot object for base graph 
(created from data_out otherwise)}

\item{alpha}{opacity of curves}
}
\value{
ggplot visualisation of the curve
}
\description{
Note / TODO: 
Function currently uses geom_path - assumed euclidean space for map :/
}
