% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde_functions.R
\name{points_in_contour_list}
\alias{points_in_contour_list}
\title{Identification of points inside a contour list}
\usage{
points_in_contour_list(cont_list, predict_mat, position = 1:2)
}
\arguments{
\item{cont_list}{list of KDE contour objects}

\item{predict_mat}{matrix with the points to be determined in terms of
position with respect to the contour.}

\item{position}{Columns position of long/lat pair. Default is 1:2}
}
\value{
Vector of binary values, 1 if the point is inside the contour, 
0 is not. Dimensionality is the same as the number of rows in predict_mat
}
\description{
Wrapper around \code{\link{points_in_contour}} for calculating whether points
are inside at least one of a series of contours
}
