% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_tc}
\alias{sample_tc}
\title{Example TC, AL032009}
\format{A data frame with 69 rows and  variables:
\describe{
  \item{date}{Date time step recorded (format "YYYYmmdd")}
  \item{time}{Time of day time step was recorded, 
  (military time, e.g. 600, 1200, 1600, 1800, 0) in UTC (Universal Time 
  Coordinate) }
  \item{record_id}{Record Identifiers, L - eye cross coastline (landfall),
  others possible - see 
  \url{http://www.aoml.noaa.gov/hrd/hurdat/newhurdat-format.pdf}}
  \item{status}{type of TC, see details}
  \item{lat}{latitude of the eye of the storm}
  \item{long}{longitude of the eye of the storm}
 ...
  \item{max_sust_wind}{Maximum sustained wind (in knots) }
  \item{min_pressure}{Minimum Pressure (in millibars) }
  \item{wind_NE_34}{34 kt wind radii maximum extent in northeastern 
  quadrant (in nautical miles) }
  \item{wind_SE_34}{... in southeastern quadrant ...}    
  \item{wind_SW_34}{... in southwestern quadrant ...}   
  \item{wind_NW_34}{... in northwestern quadrant ...}   
  \item{wind_NE_50}{50 kt wind radii maximum extent in northeastern 
  quadrant (in nautical miles) }    
  \item{wind_SE_50}{... in southeastern quadrant ...}   
  \item{wind_SW_50}{... in southwestern quadrant ...}   
  \item{wind_NW_50}{... in northwestern quadrant ...}  
  \item{wind_NE_64}{– 64 kt wind radii maximum extent in northeastern 
  quadrant (in nautical miles) }    
  \item{wind_SE_64}{... in southeastern quadrant ...}    
  \item{wind_SW_64}{... in southwestern quadrant ...}    
  \item{wind_NW_64}{... in northwestern quadrant ...}  
}}
\source{
\url{http://www.aoml.noaa.gov/hrd/hurdat/hurdat2-1851-2015-070616.txt}
\url{http://www.aoml.noaa.gov/hrd/hurdat/newhurdat-format.pdf}
\url{http://www.aoml.noaa.gov/hrd/hurdat/Data_Storm.html}
}
\usage{
sample_tc
}
\description{
A data frame of all recorded time steps of a single TC, a slightly cleaned 
version than that provided by the National Oceanic and Atmospheric 
Administration's data set. "AL032009" means that the TC was in the 
Atlantic Basin (AL), the 3rd TC that year (03), in year 2009 (2009).
}
\details{
\code{status} has possible options:
\describe{
  \item{TD}{Tropical cyclone of tropical depression intensity (\eqn{<}
  34 knots) }
  \item{TS}{Tropical cyclone of tropical storm intensity (34-63 knots) }
  \item{HU}{Tropical cyclone of hurricane intensity (\eqn{\geq} 64 knots) }
  \item{EX}{Extratropical cyclone (of any intensity) }
  \item{SD}{– Subtropical cyclone of subtropical depression intensity 
  (\eqn{<} 34 knots) }
  \item{SS}{Subtropical cyclone of subtropical storm intensity 
  (\eqn{\geq} 34 knots) }
  \item{LO}{– A low that is neither a tropical cyclone, a subtropical 
  cyclone, nor an extratropical cyclone (of any intensity) }
  \item{WV}{Tropical Wave (of any intensity)}
  \item{DB}{– Disturbance (of any intensity) }
}
}
\keyword{datasets}
