% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thirteen_point_compression.R
\name{thirteen_points_listable}
\alias{thirteen_points_listable}
\title{List of 13 point compressions}
\usage{
thirteen_points_listable(list_df, c_position = 5:6, lonlat = TRUE,
  verbose = TRUE)
}
\arguments{
\item{list_df}{list of dfs, where the lonlat points are are in the 
c_position columns}

\item{c_position}{the columns of the data frames that contain the desired 
lonlat coordinates}

\item{lonlat}{boolean logical if columns are lonlat 
(false if they are latlon)}

\item{verbose}{boolean for having a progress bar}
}
\value{
List of 13 point compression data frame for each path
}
\description{
Creates list of 13 point long/lat expression of each path
}
